Continued code cleanup.
Force all display functions to use DBGV so that all program output
is coordinated.  Perviously some output was to stdout and some to
stderr which made correlating debugging output more difficult.

Show the master to slave and slave to master values in the statistics
output.

Add display routines for delay request and delay response messages.

Add a Doxyfile and a tags target for building working cross references.
    NOTE: This requires cscope, global, doxygen and dot to be installed