Merge from maintenance branch:
Fix a math bug that occurs when we have very long (25ms or more) round
trip times when estimating the meanPathDelay and we have delay and
offset limiting in place.  If the delay is long then we wind up with a
negative delay time, and that's not good for us.