Update all files in preparation for 2.2.0 release.
Extensive changes to the ChangeLog to reflect features and bug
fixes since the 2.1.0 release.

Remove the RELEASE_NOTES as this was redundant with the
ChangeLog.

Update the INSTALL document to reflect the fact that we not longer
need most of the original OS specific options.

Add the latest contributors to the COPYRIGHT file.