Interim commit of BPF work.
Add proper constants for the BPF values.

Deal with the way in which timeouts are handled by using a smaller
snap length, based on the actual maximum packet size (300) and by
telling pcap/bpf that we want a 1ms timeout on packets.

This all solves a basic problem where the daemon was not seeing its
own delay requests in the correct order.  Normal order is

Issue (0)
Handle(0)
Issue (1)
Handle(1)

but until these changes we were seeing


Issue (0)
Issue (1)
Issue (2)
Handle(0)
Handle(1)
Handle(2)