Complete work on adding BPF as a way of receiving packets.
Add IEEE 802.3 (Ethernet) mode support.

The combination of these two features improves delay and offset measurements
and adds compatibility with clocks that support IEEE 802.3 mode.

Adding these features has required a few changes in the stuctures and
functions including:

Push the Clock UUID into the netPath structure, for use in transmitting 
IEEE 802.3 packets.

Add rtOpts to the netSend*() calls so that we have access to the
ethernet_mode boolean.

This will be the last commit on this branch.  Next the branch
will be merged into trunk as part of moving towards the 2.3 release.