Abstract R scripts into a common library for use by current and future scripts.
Remove the now redundant compare_delay.R script as it is supplanted by
being able to pass arguments such as "offset" and "delay" to the
compare.R script.

The ptplib.R file can be used to supply R command line functions that
allow the user to read log files as well as graph and compare them.