Security Extension: http://arxiv.org/abs/1603.00707
Security extension for the IEEE-1588 protocol, according to our thesis that can be found in the above link.
The above technical paper is now in discussion at the official IEEE Workgroup at it's SC.

All changes were tested and confirmed to compile and work, results are shown in the above link.